+++ commons/all/pom.xml
@@ -63,6 +63,12 @@
       </dependency>
 
       <dependency>
+         <groupId>net.openhft</groupId>
+         <artifactId>zero-allocation-hashing</artifactId>
+         <version>0.12</version>
+      </dependency>
+
+      <dependency>
          <groupId>org.infinispan</groupId>
          <artifactId>infinispan-commons-jdk11</artifactId>
          <scope>provided</scope>
